+++ client-common/src/main/java/org/realityforge/replicant/client/EntityLocator.java
@@ -1,6 +1,5 @@
 package org.realityforge.replicant.client;
 
-import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;