+++ client-common/src/main/java/org/realityforge/replicant/client/EntitySystem.java
@@ -8,8 +8,5 @@ import javax.annotation.Nonnull;
 public interface EntitySystem
 {
   @Nonnull
-  EntityRepository getRepository();
-
-  @Nonnull
   EntitySubscriptionManager getSubscriptionManager();
 }