+++ client-common/src/main/java/org/realityforge/replicant/client/transport/DataLoaderListener.java
@@ -1,7 +1,6 @@
 package org.realityforge.replicant.client.transport;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import org.realityforge.replicant.client.ChannelDescriptor;
 
 public interface DataLoaderListener