+++ client-gwt/src/main/java/org/realityforge/replicant/client/net/gwt/AbstractFrontendContextImpl.java
@@ -1,7 +1,7 @@
 package org.realityforge.replicant.client.net.gwt;
 
 import javax.annotation.Nonnull;
-import org.realityforge.replicant.client.EntityRepository;
+import org.realityforge.replicant.client.EntityLocator;
 import org.realityforge.replicant.client.EntitySubscriptionManager;
 import org.realityforge.replicant.client.runtime.AreaOfInterestService;
 import org.realityforge.replicant.client.runtime.BaseRuntimeExtension;
@@ -11,20 +11,20 @@ import org.realityforge.replicant.client.runtime.ReplicantClientSystem;
 public abstract class AbstractFrontendContextImpl
   implements BaseFrontendContext, BaseRuntimeExtension
 {
-  private final EntityRepository _repository;
+  private final EntityLocator _entityLocator;
   private final EntitySubscriptionManager _subscriptionManager;
   private final ReplicantClientSystem _replicantClientSystem;
   private final AreaOfInterestService _areaOfInterestService;
   private final ContextConverger _converger;
 
   public AbstractFrontendContextImpl( @Nonnull final ContextConverger converger,
-                                      @Nonnull final EntityRepository repository,
+                                      @Nonnull final EntityLocator entityLocator,
                                       @Nonnull final EntitySubscriptionManager subscriptionManager,
                                       @Nonnull final ReplicantClientSystem replicantClientSystem,
                                       @Nonnull final AreaOfInterestService areaOfInterestService )
   {
     _areaOfInterestService = areaOfInterestService;
-    _repository = repository;
+    _entityLocator = entityLocator;
     _subscriptionManager = subscriptionManager;
     _replicantClientSystem = replicantClientSystem;
 
@@ -72,9 +72,9 @@ public abstract class AbstractFrontendContextImpl
 
   @Nonnull
   @Override
-  public EntityRepository getRepository()
+  public EntityLocator getEntityLocator()
   {
-    return _repository;
+    return _entityLocator;
   }
 
   @Nonnull