+++ client-gwt/src/main/java/org/realityforge/replicant/sprsrc/org/realityforge/replicant/client/FilterUtil.java
@@ -3,7 +3,6 @@ package org.realityforge.replicant.client;
 import com.google.gwt.core.client.JavaScriptObject;
 import com.google.gwt.core.client.JsonUtils;
 import java.util.Objects;
-import java.util.logging.Logger;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;